requirements.txt
fastapi==0.112
uvicorn[standard]==0.30
sqlalchemy==2.0
tidb-vector==0.0.9
sentence-transformers==3.0
langchain-text-splitters==0.2
pypdf==4.3
beautifulsoup4==4.12
python-dotenv==1.0
pytest==8.3
httpx==0.27